Add files via upload
The code if for flattening a binary tree. We are recursively looking for a node with no grandchildren and both left and right child in the left subtree. Then store node->right in a temp variable and node->right becomes node->left. Then the temp variable is added at the node-right in the end and this subtree is recursively flattened.